requirements.txt
alabaster==0.7.12
awscli==1.18.14
Babel==2.8.0
botocore==1.15.14
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==5.0.3
docutils==0.16
entrypoints==0.3
flake8==3.7.9
idna==2.9
imagesize==1.2.0
imbalanced-learn==0.6.2
imblearn==0.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
packaging==20.1
pandas==1.0.1
pkg-resources==0.0.0
pyasn1==0.4.8
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.25.8
